+++ curifactory/record.py
@@ -192,18 +192,18 @@ class Record:
         reportable.record = self
         reportable.stage = self.stages[-1]
 
-        name = ""
+        qualified_name = ""
         if reportable.record.is_aggregate:
-            name = "(Aggregate)_"
+            qualified_name = "(Aggregate)_"
         if reportable.record.args is not None:
-            name += f"{reportable.record.args.name}_"
-        name += f"{reportable.stage}_"
+            qualified_name += f"{reportable.record.args.name}_"
+        qualified_name += f"{reportable.stage}_"
 
         if reportable.name is None:
-            name += str(len(self.manager.reportables))
+            qualified_name += str(len(self.manager.reportables))
         else:
-            name += reportable.name
-        reportable.name = name
+            qualified_name += reportable.name
+        reportable.qualified_name = qualified_name
 
         self.manager.reportables.append(reportable)